NO-ISSUE: address PR feedback on avatar follower
